10 01 2024
I managed to just create this csv to use as the FLAT file for the import export wizard next time I work on this problem.

I did create the table but then remembered that I both create and import through the import export wizard.

so, i ended up deleting the file that was used to create the table.