10 03 2024
So, I believe that I have found a very important thing in working with SQL.  That is that I need to get better at breaking the problem down into parts.

The first part was finding the AVG(Salary) as is so that I could get
an idea of the problem.

From here I will work next time on removing zeroes.