10 07 2024
I am moving very slowly to work through these queries.  But this is the only way I know how.  I have to take the scenic route.  It is luck to be able to take such a route.

Next time I work on this query I would like to try and do a subquery to solve to the next step.