11 12 2024
After trying a solution that previously worked.  it didn't.
I wanted to explore the DECIMAL() function but I couldn't get the query I copied from the hackerrank website.
not sure why.
I might go back and try to solve it my way.
I see I might have to do it in DECIMAL() to help with rounding?  or with the CEILING() function.
also, I gained no new ground on the CAST() function.