11 17 2024
ok, I made progress in my understanding of CASTING.

I previously thought that casting was used to change data types only but it is also used to change the precision of integers.

I believe that I am very close to the answer on my end.

that was a very confusing thing that I wasn't sure why we were using CAST() when changing data types wasn't an issue?  or at least INT to VARCHAR()